requirements.txt
PyQt5>=5.15.0
opencv-python
numpy
pyyaml
pillow
pytest
